Created initial build for anonymizer service
