+++ Anonymize/EmailAnonymizer.php
@@ -10,28 +10,39 @@
  * @link      https://www.superbrave.nl/
  */
 
-namespace SuperBrave\GdprBundle\Anonymizer;
+namespace SuperBrave\GdprBundle\Anonymize;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
+
 /**
- * Class FixedValueAnonymizer
+ * Email anonymizer class
  *
- * @package SuperBrave\GdprBundle\Anonymizer
+ * @package SuperBrave\GdprBundle\Anonymize
  */
-class FixedValueAnonymizer implements AnonymizerInterface
+class EmailAnonymizer implements AnonymizerInterface
 {
+    /**
+     * @param mixed $propertyValue  The value that has to be converted
+     * @param array $options        Options to help the anonymizer do its job
+     *
+     * @return string               The Anonymized string
+     */
     public function anonymize($propertyValue, array $options = [])
     {
         $resolver = new OptionsResolver();
         $this->configureOptions($resolver);
         $options = $resolver->resolve($options);
 
-        return $options['annotationValue'];
+        return sprintf('%s-%u@localhost', $options['fieldName'], $options['id']);
     }
 
     private function configureOptions(OptionsResolver $resolver)
     {
-        $resolver->setRequired(['annotationValue']);
+        $resolver
+            ->setRequired(['fieldName', 'id'])
+            ->setAllowedTypes('id', 'int')
+            ->setAllowedTypes('fieldName', 'string')
+        ;
     }
 }
