+++ Anonymize/FixedValueAnonymizer.php
@@ -10,39 +10,28 @@
  * @link      https://www.superbrave.nl/
  */
 
-namespace SuperBrave\GdprBundle\Anonymizer;
+namespace SuperBrave\GdprBundle\Anonymize;
 
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-
 /**
- * Email anonymizer class
+ * Class FixedValueAnonymizer
  *
- * @package SuperBrave\GdprBundle\Anonymizer
+ * @package SuperBrave\GdprBundle\Anonymize
  */
-class EmailAnonymizer implements AnonymizerInterface
+class FixedValueAnonymizer implements AnonymizerInterface
 {
-    /**
-     * @param mixed $propertyValue  The value that has to be converted
-     * @param array $options        Options to help the anonymizer do its job
-     *
-     * @return string               The Anonymized string
-     */
     public function anonymize($propertyValue, array $options = [])
     {
         $resolver = new OptionsResolver();
         $this->configureOptions($resolver);
         $options = $resolver->resolve($options);
 
-        return sprintf('%s-%u@localhost', $options['fieldName'], $options['id']);
+        return $options['annotationValue'];
     }
 
     private function configureOptions(OptionsResolver $resolver)
     {
-        $resolver
-            ->setRequired(['fieldName', 'id'])
-            ->setAllowedTypes('id', 'int')
-            ->setAllowedTypes('fieldName', 'string')
-        ;
+        $resolver->setRequired(['annotationValue']);
     }
 }
