+++ Anonymize/ObjectAnonymizeInterface.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * This file is part of the GDPR bundle.
+ *
+ * @category  Bundle
+ * @package   Gdpr
+ * @author    SuperBrave <info@superbrave.nl>
+ * @copyright 2018 SuperBrave <info@superbrave.nl>
+ * @license   https://github.com/superbrave/gdpr-bundle/blob/master/LICENSE MIT
+ * @link      https://www.superbrave.nl/
+ */
+
+namespace SuperBrave\GdprBundle\Anonymize;
+
+/**
+ * The interface defining an object is anonymizable with @see Anonymizer.
+ */
+interface ObjectAnonymizeInterface
+{
+}
