+++ Tests/Anonymizer/PropertyAnonymizerTest.php
@@ -16,9 +16,9 @@ use PHPUnit\Framework\TestCase;
 use PHPUnit_Framework_MockObject_MockObject;
 use Reflectionclass;
 use SuperBrave\GdprBundle\Annotation\Anonymize;
-use SuperBrave\GdprBundle\Anonymizer\AnonymizerCollection;
-use SuperBrave\GdprBundle\Anonymizer\AnonymizerInterface;
-use SuperBrave\GdprBundle\Anonymizer\PropertyAnonymizer;
+use SuperBrave\GdprBundle\Anonymize\AnonymizerCollection;
+use SuperBrave\GdprBundle\Anonymize\AnonymizerInterface;
+use SuperBrave\GdprBundle\Anonymize\PropertyAnonymizer;
 use SuperBrave\GdprBundle\Tests\AnnotatedMock;
 
 /**
